[#101] feat: test socket based send message work
